refactor: 마커 컨버팅 기능 클래스 분리
