+++ src/main/java/bowling/view/ResultView.java
@@ -10,7 +10,7 @@ public class ResultView {
 
     private static final int WIDTH = 6;
 
-    private static String RESULT_DELIMITER = "|";
+    private static final String RESULT_DELIMITER = "|";
 
     private ResultView() {
 
@@ -27,11 +27,12 @@ public class ResultView {
     }
 
     private static void printRow(BowlingFrames bowlingFrames, Player player) {
-        System.out.printf("|%s|", getColumn(player.getName()));
+        System.out.printf("%s%s%s", RESULT_DELIMITER, getColumn(player.getName()),
+                RESULT_DELIMITER);
         for (int i = 0; i < bowlingFrames.getSize(); i++) {
             String resultMarkers = getResultMarks(bowlingFrames.getFrame(i));
 
-            System.out.printf("%s|", getColumn(resultMarkers));
+            System.out.printf("%s%s", getColumn(resultMarkers), RESULT_DELIMITER);
         }
         System.out.println();
     }
@@ -40,62 +41,23 @@ public class ResultView {
         FallenPins firstResult = frame.getFirstTurnResult();
         FallenPins secondResult = frame.getSecondTurnResult();
 
-        String result = getResultMark(firstResult, null)
+        String result = MarkConverter.getResultMark(firstResult, null)
                 + RESULT_DELIMITER
-                + getResultMark(secondResult, firstResult);
+                + MarkConverter.getSecondMark(firstResult, secondResult);
 
         if (frame instanceof FinalFrame) {
             FinalFrame finalFrame = (FinalFrame) frame;
-            result += (RESULT_DELIMITER + getResultMark(finalFrame.getBonusTurnResult(), null));
+            result += (RESULT_DELIMITER + MarkConverter.getResultMark(finalFrame.getBonusTurnResult(), null));
         }
 
         return result.replace(RESULT_DELIMITER + ResultMark.EMPTY.getMark(),
                 ResultMark.EMPTY.getMark());
     }
 
-    private static String getResultMark(FallenPins fallenPins, FallenPins previousFallenPins) {
-        if (isEmpty(fallenPins)) {
-            return ResultMark.EMPTY.getMark();
-        }
-
-        if (isGutter(fallenPins)) {
-            return ResultMark.GUTTER.getMark();
-        }
-
-        if (isSpare(fallenPins, previousFallenPins)) {
-            return ResultMark.SPARE.getMark();
-        }
-
-        if (isStrike(fallenPins, previousFallenPins)) {
-            return ResultMark.STRIKE.getMark();
-        }
-
-        return String.valueOf(fallenPins.getCountOfPin());
-    }
-
-    private static boolean isStrike(FallenPins fallenPins, FallenPins previousFallenPins) {
-        return isEmpty(previousFallenPins)
-                && fallenPins.getCountOfPin() == FallenPins.MAX_COUNT_OF_PIN;
-    }
-
-    private static boolean isGutter(FallenPins fallenPins) {
-        return fallenPins.getCountOfPin() == 0;
-    }
-
-    private static boolean isEmpty(FallenPins fallenPins) {
-        return fallenPins == null;
-    }
-
-    private static boolean isSpare(FallenPins fallenPins, FallenPins previousFallenPins) {
-        return previousFallenPins != null &&
-                fallenPins.getCountOfPin() + previousFallenPins.getCountOfPin()
-                        == FallenPins.MAX_COUNT_OF_PIN;
-    }
-
     private static void printHeader(int frameSize) {
-        System.out.print("| NAME |");
+        System.out.printf("%s NAME %s", RESULT_DELIMITER, RESULT_DELIMITER);
         for (int i = 1; i <= frameSize; i++) {
-            System.out.printf("%s|", getColumn(String.format("%02d", i)));
+            System.out.printf("%s%s", getColumn(String.format("%02d", i)), RESULT_DELIMITER);
         }
         System.out.println();
     }
